♻️ Refactor: ReservationPanelContent 상태 로직 useReservationPanel 훅으로 분리
